standardize number of thumbnails on right-hand nav bar
Previously, "six before and six after" was the rule; now "twelve total"
is the rule, trying to draw equally from before and after, if possible,
but drawing more from one side if needed.

Also: allow an environment variable to cause the guile server to
bind to a port rather than a unix socket.